@@ -5,8 +5,8 @@
   "main": "dist/bplayer.min.js",
   "module": "src/bplayer.js",
   "scripts": {
-    "dev": "node ./build/dev-server.js",
-    "build": "node ./build/build.js",
+    "dev": "rollup -c ./config/rollup.dev.js -w",
+    "build": "NODE_ENV=production rollup -c ./config/rollup.prod.js",
     "lint": "eslint --ext .js src"
   },
   "repository": {
@@ -27,19 +27,20 @@
   },
   "homepage": "http://bplayer.js.org",
   "devDependencies": {
-    "browser-sync": "^2.17.5",
-    "eslint": "^3.9.1",
-    "node-watch": "^0.4.1",
-    "rollup": "^0.36.3",
+    "chalk": "^2.1.0",
+    "chokidar": "^1.7.0",
+    "eslint": "^4.5.0",
+    "rollup": "^0.48.2",
+    "rollup-plugin-browsersync": "^0.2.4",
     "rollup-plugin-buble": "^0.15.0",
-    "rollup-plugin-commonjs": "^5.0.5",
-    "rollup-plugin-eslint": "^3.0.0",
+    "rollup-plugin-commonjs": "^8.2.0",
+    "rollup-plugin-eslint": "^4.0.0",
     "rollup-plugin-html": "^0.2.1",
-    "rollup-plugin-node-resolve": "^2.0.0",
-    "rollup-plugin-postcss": "^0.2.0",
-    "rollup-plugin-progress": "^0.1.0",
+    "rollup-plugin-json": "^2.3.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-postcss": "^0.5.4",
+    "rollup-plugin-progress": "^0.4.0",
     "rollup-plugin-replace": "^1.1.1",
-    "rollup-plugin-uglify": "^1.0.1",
-    "shelljs": "^0.7.5"
+    "rollup-plugin-uglify": "^2.0.1"
   }
 }
